// frontend/src/app/api-gen/models/params-multi-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { MultiResultInfo } from '../models/multi-result-info';
export interface ParamsMultiResponse {
  create_secret: boolean;
  dry_run: boolean;
  err_count: number;
  name: string;
  ok_count: number;
  results: Array<MultiResultInfo>;
  value_type: string;
}